Improve variant search query time
This improves both the count query for pagination and the query itself
when paginated for large amounts of results, by avoiding both sorting on
the serialized column, and by deferring pulling out the serialized data
until after the limiting is finished. It does this by doing the "work"
in a subquery that then pulls out only the IDs, and then pulling out the
serialized data mapped to these IDs afterward.